GEODE-8537: Solve LRU eviction mem increase
 - LRUEntryProperties has been put into a separate file.
 - LRUList has been entirely refactored. It now uses STL's deque as base
   implementation. Implements the following methods:
   * push
   * pop
   * remove
   * move_to_end
   * size
   * clear
 - LRUList has been renamed to LRUQueue.
 - Refactored LRUEntriesMap to integrate with the new implementation of
   LRUQueue.
 - Added UTs for LRUQueue.
 - Added a mock for MapEntryImpl.
 - Removed testLRUList from old integration tests as has been replaced by the UT
   TS.
 - EvictionController and EvictionThread was causing an inter-lock and also
   there was a race-condition which made LRU eviction not to be accurate.
 - In order to solve EvictionController malfunction the following things have
   been changed:
   * Now evictions are made by the EvictionController and the EvictionThread has
     been removed as in addition to the inter-lock, it was causing some
     unnecessary overhead.
   * EvictionController queue has been replaced by the heap size itself as an
     atomic variable. So now, eviction happens directly whenever heap size
     overflows the established limit, without having to sum up the deltas on the
     queue.
 - Additionally heap size counter in the LRUEntriesMap has been replaced by an
   atomic counter, getting rid therefore of the mutex.
 - Added a new integrations TS to verify LRU eviction functionality.
 - After implementing the new LRUQueue I noticed that
   testRegionMap.TestRecentlyUsedBit was failing and I noticed that the test was
   wrongly implemented.
 - Having set LRU entries limit to 10 and getting key "15" you could not expect
   that after putting keys [20, 35], key "15" would be still there.
 - Also the test has been replaced by one new integration test called
   LRUEvictionTest.verifyLruOrderGet
 - Used mixed case in the EvictionController.
 - Removed debugging trace which was left during development.
 - After discussing about the nature of LRUQueue, it seems that
   std::mutex is a better alternative to make LRUQueue thread-safe, so
   it's now the mutex used.

Status: Incomplete
Jira: ADPPRG-37777

SQUASH

+++ cppcache/src/LocalRegion.hpp
@@ -394,7 +394,7 @@ class APACHE_GEODE_EXPORT LocalRegion : public RegionInternal {
                          const std::string& factoryFuncName) override;
   CacheImpl* getCacheImpl() const override;
 
-  void evict(int32_t percentage) override;
+  void evict(float percentage) override;
 
   virtual void acquireGlobals(bool isFailover);
